Consolidates logic by extracting IteratorPeeking out of RelPeeking
