+++ partiql-eval/src/main/kotlin/org/partiql/eval/internal/helpers/IteratorPeeking.kt
@@ -1,0 +1,36 @@
+package org.partiql.eval.internal.helpers
+
+/**
+ * For [Iterator]s that MUST materialize data in order to execute [hasNext], this abstract class caches the
+ * result of [peek] to implement both [hasNext] and [next].
+ *
+ * With this implementation, invoking hasNext() multiple times will not iterate unnecessarily. Invoking next() without
+ * invoking hasNext() is allowed -- however, it is highly recommended to avoid doing so.
+ */
+abstract class IteratorPeeking<T> : Iterator<T> {
+
+    private var _next: T? = null
+
+    /**
+     * @return NULL when there is not another [T] to be produced. Returns a [T] when able to.
+     *
+     * @see IteratorPeeking
+     */
+    abstract fun peek(): T?
+
+    override fun hasNext(): Boolean {
+        if (_next != null) {
+            return true
+        }
+        this._next = peek()
+        return this._next != null
+    }
+
+    override fun next(): T {
+        val next = _next
+            ?: peek()
+            ?: error("There was no more elements, however, next() was called. Please use hasNext() beforehand.")
+        this._next = null
+        return next
+    }
+}
